Phase-5/backend/test_advanced_features.py
"""
Comprehensive tests for Phase 5 advanced features
Tests all 7 new features plus event-driven architecture
"""

import pytest
import asyncio
from datetime import datetime, timedelta
from sqlmodel import Session, select
from unittest.mock import Mock, patch

from models import Task, User, Tag, TaskTagLink
from main import app
from dapr_integration import dapr_client
from event_handlers import event_handler


@pytest.fixture
def test_db_session():
    """Create a test database session"""
    from database import engine
    with Session(engine) as session:
        yield session


@pytest.mark.asyncio
async def test_recurring_task_creation(test_db_session):
    """Test creating a recurring task"""
    # Create user
    user = User(id="test_user", email="test@example.com")
    test_db_session.add(user)
    test_db_session.commit()

    # Create a recurring task
    from models import TaskCreate
    task_data = TaskCreate(
        title="Recurring Task",
        description="A task that repeats",
        recurrence_pattern="daily"
    )

    task = Task(
        title=task_data.title,
        description=task_data.description,
        recurrence_pattern=task_data.recurrence_pattern,
        owner_id=user.id
    )

    test_db_session.add(task)
    test_db_session.commit()
    test_db_session.refresh(task)

    assert task.recurrence_pattern == "daily"
    assert task.title == "Recurring Task"


def test_priority_levels(test_db_session):
    """Test priority level functionality"""
    user = User(id="test_user_2", email="test2@example.com")
    test_db_session.add(user)
    test_db_session.commit()

    # Test all priority levels
    priorities = ["high", "medium", "low"]

    for priority in priorities:
        task = Task(
            title=f"Task with {priority} priority",
            priority=priority,
            owner_id=user.id
        )
        test_db_session.add(task)

    test_db_session.commit()

    # Verify all priorities were saved
    tasks = test_db_session.exec(select(Task).where(Task.owner_id == user.id)).all()
    saved_priorities = [task.priority for task in tasks]

    for priority in priorities:
        assert priority in saved_priorities


def test_due_date_functionality(test_db_session):
    """Test due date functionality"""
    user = User(id="test_user_3", email="test3@example.com")
    test_db_session.add(user)
    test_db_session.commit()

    future_date = datetime.now() + timedelta(days=7)

    task = Task(
        title="Task with due date",
        due_date=future_date,
        owner_id=user.id
    )

    test_db_session.add(task)
    test_db_session.commit()
    test_db_session.refresh(task)

    assert task.due_date == future_date


def test_tags_functionality(test_db_session):
    """Test tags functionality"""
    user = User(id="test_user_4", email="test4@example.com")
    test_db_session.add(user)
    test_db_session.commit()

    # Create task
    task = Task(title="Tagged Task", owner_id=user.id)
    test_db_session.add(task)
    test_db_session.commit()
    test_db_session.refresh(task)

    # Create tags
    tag1 = Tag(name="work", user_id=user.id)
    tag2 = Tag(name="urgent", user_id=user.id)
    test_db_session.add(tag1)
    test_db_session.add(tag2)
    test_db_session.commit()

    # Link task to tags
    link1 = TaskTagLink(task_id=task.id, tag_id=tag1.id)
    link2 = TaskTagLink(task_id=task.id, tag_id=tag2.id)
    test_db_session.add(link1)
    test_db_session.add(link2)
    test_db_session.commit()

    # Verify tags are linked
    task_links = test_db_session.exec(
        select(TaskTagLink).where(TaskTagLink.task_id == task.id)
    ).all()

    assert len(task_links) == 2


def test_full_text_search_simulation():
    """Test full-text search functionality (simulated)"""
    # This would normally test the search endpoint
    # For now, we'll verify the search fields exist in the model
    from models import Task

    # Check that Task model has searchable fields
    task = Task(
        title="Searchable Task Title",
        description="This is a searchable description",
        owner_id="test_owner"
    )

    # Verify all required fields exist
    assert hasattr(task, 'title')
    assert hasattr(task, 'description')
    assert task.title == "Searchable Task Title"
    assert task.description == "This is a searchable description"


def test_advanced_filters_simulation():
    """Test advanced filtering functionality (simulated)"""
    # This would normally test the filtering endpoints
    # For now, we'll verify the filterable fields exist in the model
    from models import Task

    task = Task(
        title="Filterable Task",
        priority="high",
        is_completed=False,
        owner_id="test_owner"
    )

    # Verify filterable attributes exist
    assert hasattr(task, 'priority')
    assert hasattr(task, 'is_completed')
    assert task.priority == "high"
    assert task.is_completed == False


def test_sorting_functionality_simulation():
    """Test sorting functionality (simulated)"""
    # This would normally test the sorting endpoints
    # For now, we'll verify the sortable fields exist in the model
    from models import Task

    task1 = Task(
        title="A Task",
        priority="low",
        created_at=datetime.now(),
        owner_id="test_owner"
    )

    task2 = Task(
        title="Z Task",
        priority="high",
        created_at=datetime.now() + timedelta(minutes=1),
        owner_id="test_owner"
    )

    # Verify sortable attributes exist
    assert hasattr(task1, 'title')
    assert hasattr(task1, 'priority')
    assert hasattr(task1, 'created_at')
    assert task1.title < task2.title  # Alphabetical sort
    assert task1.priority != task2.priority  # Different priorities


@pytest.mark.asyncio
async def test_dapr_event_publishing():
    """Test Dapr event publishing functionality"""
    # Mock the Dapr client
    with patch.object(dapr_client, 'client') as mock_dapr_client:
        mock_dapr_client.publish_event = Mock()

        # Test publishing a task.created event
        event_data = {
            "task_id": 1,
            "user_id": "test_user",
            "title": "Test Task",
            "priority": "medium"
        }

        # Initialize the client
        dapr_client.initialize()

        # Publish event
        dapr_client.publish_task_event("task.created", event_data)

        # Verify event was published
        mock_dapr_client.publish_event.assert_called_once()


@pytest.mark.asyncio
async def test_event_handling():
    """Test event handling functionality"""
    # Mock the database session
    with patch.object(event_handler, 'session') as mock_session:
        mock_session.get = Mock(return_value=Mock())

        # Test handling a task.created event
        event_data = {
            "task_id": 1,
            "user_id": "test_user",
            "title": "Test Task",
            "priority": "medium"
        }

        # Handle the event
        await event_handler.handle_task_created(event_data)

        # Verify the event was handled
        # (In a real test, we'd verify specific behaviors)


if __name__ == "__main__":
    pytest.main([__file__])