Phase-5/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlmodel==0.0.22
psycopg2-binary>=2.9.11
python-jose[cryptography]==3.3.0
pydantic-settings==2.7.0
python-dotenv==1.0.1
SQLAlchemy==2.0.36
cohere==5.20.0
requests==2.32.3
asyncio==3.4.3
httpx==0.27.0
dapr==1.12.0